requirements.txt
attrs==17.3.0
Automat==0.6.0
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
enum-compat==0.0.2
enum34==1.1.6
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
influxdb==4.1.1
netifaces==0.10.6
nodeenv==1.2.0
olefile==0.44
Pillow==4.3.0
pyserial==3.4
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
Twisted==17.9.0
urllib3==1.22
zeroconf==0.19.1
zope.interface==4.4.3
